test: Primalityのテストの修正強化
